backend/pyproject.toml
[tool.poetry]
name = "askcandy"
version = "0.1.0"
description = ""
authors = ["noname2048 <sungwook.csw@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.8"
Django = "^3.2"
black = "^20.8b1"
environs = "^9.3.2"
Pillow = "^8.2.0"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.7.0"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
